+++ source/YoutubeDLSharp/Options/Enums.cs
@@ -1,0 +1,34 @@
+﻿namespace YoutubeDLSharp.Options;
+
+public enum DownloadMergeFormat
+{
+    Unspecified,
+    Mp4,
+    Mkv,
+    Ogg,
+    Webm,
+    Flv,
+}
+
+public enum AudioConversionFormat
+{
+    Best,
+    Aac,
+    Flac,
+    Mp3,
+    M4a,
+    Opus,
+    Vorbis,
+    Wav,
+}
+
+public enum VideoRecodeFormat
+{
+    None,
+    Mp4,
+    Mkv,
+    Ogg,
+    Webm,
+    Flv,
+    Avi,
+}